// Desktop/Eventmeet/apps/web/src/types/notification.ts
export type NotificationType =
  | 'CONNECTION_REQUEST'
  | 'CONNECTION_ACCEPTED'
  | 'GROUP_INVITE'
  | 'BOOKING_INVITE'
  | 'QUEUE_UPDATED'
  | 'SEAT_ASSIGNED'
  | 'BOOKING_CONFIRMED'
  | 'BOOKING_CANCELLED'
  | 'PAYMENT_FAILED'
  | 'EVENT_CANCELLED'
  | 'RATING_REMINDER'
  | 'CREDITS_RECHARGED'
  | 'QUICK_CONNECT_REQUEST'